+++ packages/react-liveness/src/components/FaceLivenessDetector/service/machine/index.ts
@@ -49,6 +49,7 @@ import { STATIC_VIDEO_CONSTRAINTS } from '../../StartLiveness/helpers';
 import { WS_CLOSURE_CODE } from '../utils/constants';
 
 export const MIN_FACE_MATCH_TIME = 500;
+const DEFAULT_FACE_FIT_TIMEOUT = 7000;
 
 // timer metrics variables
 let faceDetectedTimestamp: number;
@@ -675,7 +676,13 @@ export const livenessMachine = createMachine<LivenessContext, LivenessEvent>(
       sendTimeoutAfterOvalMatchDelay: actions.send(
         { type: 'TIMEOUT' },
         {
-          delay: 7000,
+          delay: (context) => {
+            return (
+              context.serverSessionInformation?.Challenge
+                ?.FaceMovementAndLightChallenge?.ChallengeConfig
+                ?.OvalFitTimeout || DEFAULT_FACE_FIT_TIMEOUT
+            );
+          },
           id: 'ovalMatchTimeout',
         }
       ),