+++ packages/react/src/primitives/Select/Select.tsx
@@ -4,7 +4,7 @@ import classNames from 'classnames';
 import { classNameModifier, classNameModifierByFlag } from '../shared/utils';
 import { ComponentClassNames } from '../shared/constants';
 import { Flex } from '../Flex';
-import { IconExpandMore } from '../Icon/internal';
+import { IconExpandMore, useIcons } from '../Icon';
 import { ForwardRefPrimitive, Primitive } from '../types';
 import { BaseSelectProps, SelectProps } from '../types/select';
 import { View } from '../View';
@@ -18,7 +18,7 @@ const SelectPrimitive: Primitive<SelectProps, 'select'> = (
     value,
     defaultValue,
     hasError,
-    icon = <IconExpandMore />,
+    icon,
     iconColor,
     children,
     placeholder,
@@ -40,6 +40,7 @@ const SelectPrimitive: Primitive<SelectProps, 'select'> = (
     classNameModifierByFlag(ComponentClassNames.Select, 'error', hasError),
     className
   );
+  const icons = useIcons('select');
 
   return (
     <View className={ComponentClassNames.SelectWrapper}>
@@ -71,7 +72,7 @@ const SelectPrimitive: Primitive<SelectProps, 'select'> = (
         )}
         color={iconColor}
       >
-        {icon}
+        {icon ?? icons?.expand ?? <IconExpandMore />}
       </Flex>
     </View>
   );