+++ packages/ui/src/theme/tokens/components/storagemanager.ts
@@ -2,8 +2,11 @@ import { DesignTokenProperties, OutputVariantKey } from '../types/designToken';
 import { fileuploader, FileUploaderTokens } from './fileUploader';
 
 export interface StorageManagerTokens<OutputType extends OutputVariantKey>
-  extends Omit<FileUploaderTokens<OutputType>, 'body' | 'footer'> {
+  extends Omit<FileUploaderTokens<OutputType>, 'body' | 'previewer'> {
   filelist?: DesignTokenProperties<'flexDirection' | 'gap'>;
+  previewer?: {
+    footer?: DesignTokenProperties<'justifyContent', OutputType>;
+  };
 }
 
 export const storagemanager: Required<StorageManagerTokens<'default'>> = {
@@ -11,5 +14,11 @@ export const storagemanager: Required<StorageManagerTokens<'default'>> = {
   filelist: {
     flexDirection: { value: 'column' },
     gap: { value: '{space.small}' },
+  },
+  previewer: {
+    ...fileuploader.previewer,
+    footer: {
+      justifyContent: { value: 'flex-end' },
+    },
   },
 };